+++ src/AbstractTypedArray.php
@@ -87,7 +87,7 @@ abstract class AbstractTypedArray implements TypedArrayInterface, ReadOnlyInterf
             $offsetWasNull = false;
         }
 
-        if ($this->canAddValue($offset, $value)) {
+        if ($this->canAddValue($value)) {
             if ($offsetWasNull) {
                 $this->nextIntKey++;
             } elseif (is_int($offset) && $offset >= $this->nextIntKey) {
@@ -194,7 +194,7 @@ abstract class AbstractTypedArray implements TypedArrayInterface, ReadOnlyInterf
         return is_null($value) ? 'NULL' : (string) $value;
     }
 
-    protected function canAddValue(mixed $offset, mixed $value): bool
+    protected function canAddValue(mixed $value): bool
     {
         $return = true;
 
@@ -209,7 +209,7 @@ abstract class AbstractTypedArray implements TypedArrayInterface, ReadOnlyInterf
                 if ($this->isSameValues($value, $internalValue)) {
                     if ($this->getValueAlreadyExistMode() === ValueAlreadyExistsModeEnum::EXCEPTION) {
                         throw new ValueAlreadyExistException(
-                            'Value "' . $this->castValueToString($value) . '" already exist.'
+                            'Value "' . $this->castValueToString($value) . '" already exists.'
                         );
                     }