+++ src/EnumArray/AbstractEnumArray.php
@@ -55,7 +55,7 @@ abstract class AbstractEnumArray extends AbstractTypedArray
         return $firstValue === $secondValue;
     }
 
-    protected function canAddValue(mixed $offset, mixed $value): bool
+    protected function canAddValue(mixed $value): bool
     {
         if (
             is_object($value) === false
@@ -64,7 +64,7 @@ abstract class AbstractEnumArray extends AbstractTypedArray
             throw new InvalidTypeException('$value should be an instance of ' . $this->getClassName() . '.');
         }
 
-        return parent::canAddValue($offset, $value);
+        return parent::canAddValue($value);
     }
 
     protected function castValueToString(mixed $value): string