+++ src/ObjectArray/AbstractObjectArray.php
@@ -30,10 +30,10 @@ abstract class AbstractObjectArray extends AbstractTypedArray
         return '$value should be an instance of ' . $this->getClassName() . '.';
     }
 
-    protected function canAddValue(mixed $offset, mixed $value): bool
+    protected function canAddValue(mixed $value): bool
     {
         $this->assertInstanceOf($value);
 
-        return parent::canAddValue($offset, $value);
+        return parent::canAddValue($value);
     }
 }