+++ src/ScalarArray/StringNullableArray.php
@@ -43,12 +43,12 @@ class StringNullableArray extends AbstractTypedArray implements ScalarArrayInter
         return parent::toArray();
     }
 
-    protected function canAddValue(mixed $offset, mixed $value): bool
+    protected function canAddValue(mixed $value): bool
     {
         if (is_null($value) === false && is_string($value) === false) {
             throw new InvalidTypeException('$value should be of type string or null.');
         }
 
-        return parent::canAddValue($offset, $value);
+        return parent::canAddValue($value);
     }
 }